test(app): express Explorer panel regression as user steps
