+++ packages/app/e2e/browser/explorer-plugin-menu.spec.ts
@@ -1,187 +1,23 @@
-import { mkdtemp, rm, writeFile } from "node:fs/promises";
-import { tmpdir } from "node:os";
-import path from "node:path";
-import { expect, test, type Page } from "../support/fixtures";
-import { gotoWorkspace } from "../support/helpers/launcher";
-import { buildAgentRoute } from "../support/helpers/mock-agent";
-import { connectNewWorkspaceDaemonClient } from "../support/helpers/new-workspace";
-import { pluginRequirements } from "../support/helpers/plugin-fixture";
-import { seedWorkspace } from "../support/helpers/seed-client";
-import { ensureExplorerSidebar } from "../support/helpers/workspace-tabs";
-
-async function writePanelPlugin(id: string, title: string) {
-  const directory = await mkdtemp(path.join(tmpdir(), "paseo-explorer-menu-"));
-  await writeFile(
-    path.join(directory, "paseo-plugin.json"),
-    JSON.stringify({ id, requirements: pluginRequirements }),
-  );
-  await writeFile(
-    path.join(directory, "index.client.tsx"),
-    `
-import React from "react";
-import { Text } from "react-native";
-export default function contribute(client) {
-  client.addWorkspacePanel({ id: "review", title: ${JSON.stringify(title)}, icon: "Blocks", context: "workspace", locations: ["workspace", "explorer"], Component: ({ workspaceId }) => <Text>{${JSON.stringify(title)} + " workspace " + workspaceId}</Text> });
-  client.addWorkspacePanel({ id: "summary", title: ${JSON.stringify(`${title} summary`)}, icon: "Blocks", context: "workspace", locations: ["explorer"], Component: () => <Text>Summary content</Text> });
-  client.addWorkspacePanel({ id: "main", title: ${JSON.stringify(`${title} main only`)}, icon: "Blocks", context: "workspace", Component: () => null });
-  client.addWorkspacePanel({ id: "agent", title: ${JSON.stringify(`${title} agent`)}, icon: "Blocks", context: "agent", locations: ["explorer"], Component: () => null });
-  return () => {};
-}`,
-  );
-  return directory;
-}
-
-async function openExplorerMenu(page: Page) {
-  const explorer = await ensureExplorerSidebar(page);
-  // The rail's empty background has no accessible role; tab menus are separate.
-  await explorer.getByTestId("explorer-sidebar-tab-rail").click({
-    button: "right",
-    position: { x: 20, y: 2 },
-  });
-  const menu = page.getByTestId("explorer-sidebar-tab-configuration");
-  await expect(menu).toBeVisible();
-  await expect(menu).toHaveCSS("opacity", "1");
-  return menu;
-}
-
-async function toggleExplorerView(page: Page, name: string) {
-  const menu = await openExplorerMenu(page);
-  await menu.getByRole("menuitem", { name, exact: true }).click();
-  await expect(menu).not.toBeVisible();
-}
+import { test } from "../support/fixtures";
+import {
+  withExplorerPanelPlugins,
+  expectWorkspacePanelsInExplorerLauncher,
+  openWorkspacePanelFromExplorerMenu,
+  expectIndependentExplorerPanelToggles,
+  expectExplorerPanelWithFocusedAgent,
+} from "../support/helpers/explorer-plugin-menu";
 
 test("Explorer toggles workspace panels independently without inheriting agent context", async ({
   page,
 }, testInfo) => {
-  const client = await connectNewWorkspaceDaemonClient({ ownProjects: false });
-  const previousConfig = await client.getDaemonConfig();
-  const workspace = await seedWorkspace({ repoPrefix: "explorer-plugin-menu-" });
-  const first = await writePanelPlugin("explorer-review", "Review");
-  const second = await writePanelPlugin("explorer-other", "Other review");
-  try {
-    await client.patchDaemonConfig({ pluginsEnabled: true });
-    await client.installDirectoryPlugin(first);
-    await client.installDirectoryPlugin(second);
-    await page.setViewportSize({ width: 1280, height: 900 });
-    await gotoWorkspace(page, workspace.workspaceId);
-
-    await test.step("the New Tab launcher offers compatible workspace panels", async () => {
-      const menu = await openExplorerMenu(page);
-      await menu.getByRole("menuitem", { name: "New tab", exact: true }).click();
-      const launcher = page.getByTestId("workspace-explorer-sidebar");
-      await expect(launcher.getByRole("button", { name: "Review", exact: true })).toBeVisible();
-      await expect(launcher.getByRole("button", { name: "Review agent", exact: true })).toHaveCount(
-        0,
-      );
-      await expect(
-        launcher.getByRole("button", { name: "Review main only", exact: true }),
-      ).toHaveCount(0);
-    });
-
-    await test.step("the rail menu offers the same workspace panels", async () => {
-      const menu = await openExplorerMenu(page);
-      await testInfo.attach("explorer-panel-menu", {
-        body: await page.screenshot({ path: testInfo.outputPath("explorer-panel-menu.png") }),
-        contentType: "image/png",
-      });
-      await expect(menu.getByRole("menuitem")).toHaveText([
-        "New tab",
-        "Changes",
-        "Files",
-        "Other review",
-        "Other review summary",
-        "Review",
-        "Review summary",
-      ]);
-      await expect(menu.getByRole("menuitem", { name: "Review", exact: true })).toHaveAttribute(
-        "aria-checked",
-        "false",
-      );
-      await menu.getByRole("menuitem", { name: "Review", exact: true }).click();
-      await expect(
-        page.getByText(`Review workspace ${workspace.workspaceId}`, { exact: true }),
-      ).toBeVisible();
-    });
-
-    await test.step("different panels and plugins keep independent selection and closing", async () => {
-      await toggleExplorerView(page, "Other review");
-      await expect(
-        page.getByText(`Other review workspace ${workspace.workspaceId}`, { exact: true }),
-      ).toBeVisible();
-      await toggleExplorerView(page, "Review summary");
-      const menu = await openExplorerMenu(page);
-      await expect(menu.getByRole("menuitem", { name: "Review", exact: true })).toHaveAttribute(
-        "aria-checked",
-        "true",
-      );
-      await expect(
-        menu.getByRole("menuitem", { name: "Other review", exact: true }),
-      ).toHaveAttribute("aria-checked", "true");
-      await expect(
-        menu.getByRole("menuitem", { name: "Review summary", exact: true }),
-      ).toHaveAttribute("aria-checked", "true");
-      await menu.getByRole("menuitem", { name: "Review", exact: true }).click();
-      const updated = await openExplorerMenu(page);
-      await expect(updated.getByRole("menuitem", { name: "Review", exact: true })).toHaveAttribute(
-        "aria-checked",
-        "false",
-      );
-      await expect(
-        updated.getByRole("menuitem", { name: "Other review", exact: true }),
-      ).toHaveAttribute("aria-checked", "true");
-      await expect(
-        updated.getByRole("menuitem", { name: "Review summary", exact: true }),
-      ).toHaveAttribute("aria-checked", "true");
-      await page.keyboard.press("Escape");
-      await toggleExplorerView(page, "Files");
-      await toggleExplorerView(page, "Files");
-      await expect(
-        page
-          .getByTestId("explorer-sidebar-tab-rail")
-          .getByRole("button", { name: "Browse workspace files", exact: true }),
-      ).toHaveCount(1);
-    });
-
-    await test.step("focusing an agent does not add agent panels to Explorer", async () => {
-      const agent = await workspace.client.createAgent({
-        provider: "mock",
-        cwd: workspace.repoPath,
-        workspaceId: workspace.workspaceId,
-        title: "Focused agent",
-        model: "ten-second-stream",
-        modeId: "load-test",
-      });
-      await page.goto(buildAgentRoute(workspace.workspaceId, agent.id));
-      await expect(
-        page.getByTestId(`workspace-tab-agent_${agent.id}`).filter({ visible: true }),
-      ).toHaveAttribute("aria-selected", "true");
-      const menu = await openExplorerMenu(page);
-      await expect(menu.getByRole("menuitem", { name: "Review agent", exact: true })).toHaveCount(
-        0,
-      );
-      await expect(
-        menu.getByRole("menuitem", { name: "Other review agent", exact: true }),
-      ).toHaveCount(0);
-      await menu.getByRole("menuitem", { name: "Review", exact: true }).click();
-      await expect(
-        page.getByText(`Review workspace ${workspace.workspaceId}`, { exact: true }),
-      ).toBeVisible();
-      await testInfo.attach("workspace-panel-with-focused-agent", {
-        body: await page.screenshot({
-          path: testInfo.outputPath("workspace-panel-with-focused-agent.png"),
-        }),
-        contentType: "image/png",
-      });
-    });
-  } finally {
-    await client.removePlugin("explorer-review");
-    await client.removePlugin("explorer-other");
-    await client.patchDaemonConfig({
-      pluginsEnabled: previousConfig.config.pluginsEnabled ?? false,
-    });
-    await client.close();
-    await workspace.cleanup();
-    await rm(first, { recursive: true, force: true });
-    await rm(second, { recursive: true, force: true });
-  }
+  await withExplorerPanelPlugins(page, async (workspace) => {
+    await test.step("the New Tab launcher offers compatible workspace panels", () =>
+      expectWorkspacePanelsInExplorerLauncher(page));
+    await test.step("the rail menu offers the same workspace panels", () =>
+      openWorkspacePanelFromExplorerMenu(page, workspace, testInfo));
+    await test.step("different panels and plugins keep independent selection and closing", () =>
+      expectIndependentExplorerPanelToggles(page, workspace));
+    await test.step("focusing an agent does not add agent panels to Explorer", () =>
+      expectExplorerPanelWithFocusedAgent(page, workspace, testInfo));
+  });
 });
